chore: edited install-frontend for windows

+++ install-frontend.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 
 const { execSync } = require('child_process')
+const os = require('os')
 
 const frontendGitUrl = 'https://github.com/chingu-voyage4/Bears-Team-6.git'
 const dirName = 'Bears-Team-6'
@@ -14,10 +15,18 @@ const execAndPrint = (command) => print(execSync(command))
 
 const installFrontendScript = () => {
   printInfo('Removing previous frontend repo')
-  execAndPrint('rm -rf ./Bears-Team-6')
+  if (os.platform() === 'win32') {
+    execAndPrint('del "Bears-Team-6"')
+  } else {
+    execAndPrint('rm -rf ./Bears-Team-6')
+  }
 
   printInfo('Removing previous frontend build')
-  execAndPrint('rm -rf ./static')
+  if (os.platform() === 'win32') {
+    execAndPrint('del "static"')
+  } else {
+    execAndPrint('rm -rf ./static')
+  }
 
   printInfo('Cloning fresh frontend repo')
   execAndPrint(`git clone -b ${branchName} ${frontendGitUrl}`)
@@ -32,10 +41,18 @@ const installFrontendScript = () => {
   // execAndPrint(`cd ${dirName} && npm run build && cd ..`)
 
   printInfo('Moving frontend')
-  execAndPrint(`mv ./${dirName}/build ./static`)
+  if (os.platform() === 'win32') {
+    execAndPrint(`move "${dirName}/build" "static"`)
+  } else {
+    execAndPrint(`mv ./${dirName}/build ./static`)
+  }
 
   printInfo('Removing frontend repo')
-  execAndPrint('rm -rf ./Bears-Team-6')
+  if (os.platform() === 'win32') {
+    execAndPrint('del "Bears-Team-6"')
+  } else {
+    execAndPrint('rm -rf ./Bears-Team-6')
+  }
 }
 
 try {
